requirements.txt
Django<6.0,>=5.0
django-allauth
django-ckeditor-5==0.2.12
python-dateutil==2.9.0.post0
requests==2.32.5
django-extensions
polib
python-dotenv
resend
weasyprint
openai
stripe
